Extract init function, so that obj file can be used elsewhere
