+++ types/func_darwin.go
@@ -2,26 +2,20 @@ package types
 
 import (
 	"debug/macho"
-	"os"
+	"fmt"
 )
 
-func init() {
-	f, err := macho.Open(os.Args[0])
-	if err != nil {
-		panic("cannot read Mach-O binary: " + err.Error())
-	}
-
+func initMachOFunctionTables(f *macho.File) error {
 	pclntab := f.Section("__gopclntab")
 	pclntabData, err := readAll(pclntab, pclntab.Size)
 	if err != nil {
-		panic("cannot read pclntab: " + err.Error())
+		return fmt.Errorf("cannot read pclntab: %w", err)
 	}
-
 	symtab := f.Section("__gosymtab")
 	symtabData, err := readAll(symtab, symtab.Size)
 	if err != nil {
-		panic("cannot read symtab: " + err.Error())
+		return fmt.Errorf("cannot read symtab: %w", err)
 	}
-
 	initFunctionTables(pclntabData, symtabData)
+	return nil
 }
